MAJOR OVERHAUL!
Constructor function used to create each shelf: read, reading and to-read
Each shelf is now an object of the Bookshelf prototype, with access to the following properties/methods:
.table: refers to the corresponding html table element for this shelf
.getBooks: retrieve array of books for this shelf from local storage
.setBooks: store updated array of books for this shelf into local storage
Previous properties and methods eg addBook and renderTable which are bound to the specific shelf

All Bookshelf instances are properties of the library object, which is created via IIFE

Abstraction of methods: Library object has access to methods like addBook() and renderTable() which will call the
corresponding addBook() and renderTable() methods within a targeted shelf of the library. This shelf is
targeted by being passed as an argument into the library's addBook() and renderTable() methods.