Filtering and sorting now work seamlessly with removeBook()
Solution: Whenever getBooks() is called, the index of each book within the stored array (in localStorage) is saved as a propetty of the book object. Therefore, even when filterShelf() and sortShelf() alter the RENDERED indices of the books rendered in the table, the original indices of the books are preserved. This allows us to accurately identify the book to be removed from the stored array via its index in the stored array, which has not changed.